Real_Estate_Management_System_Django/requirements.txt
Django==3.2
Pillow==8.2.0
psycopg2==2.8.6
psycopg2-binary==2.8.6
pytz==2021.1

form~=0.0.1
asgiref==3.6.0
automium==0.2.6
automium-web==0.1.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
contourpy==1.0.7
cryptography==40.0.1
cycler==0.11.0
defusedxml==0.7.1
Django==3.2
django-allauth==0.54.0
django-otp==1.1.6
django-recaptcha==3.0.0
Flask==2.2.3
fonttools==4.39.3
form==0.0.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.4
legacy==0.1.6
MarkupSafe==2.1.2
matplotlib==3.7.1
mysqlclient==2.1.1
numpy==1.24.2
oauthlib==3.2.2
packaging==23.0
Pillow==9.4.0
pycparser==2.21
PyJWT==2.6.0
pyotp==2.8.0
pyparsing==3.0.9
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.1
requests==2.28.2
requests-oauthlib==1.3.1
scipy==1.10.1
six==1.16.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.4.3
tzdata==2023.3
urllib3==1.26.15
views==0.3
Werkzeug==2.2.3
